Move dyson setup into its own file and call it in the pa11y and browser tests
